Rename installed files from sqlite3 to sqlite3-cipher
sqlcipher has the annoying habit of overwriting an already installed
version of sqlite3, esepcially if it is installed to the same prefix
where a normal SQLite 3 would be installed. Unfortunately this causes
issues because programs require certain versions of sqlite3 and don't
work with the version that is installed with sqlcipher.

This renames a couple of key files, sets up better defaults in the
configure file and allows a sqlite3 install to co-exist happily with a
sqlcipher install.

The following files are installed by this modification:

    ./bin
    ./bin/sqlite3-cipher
    ./include
    ./include/sqlite3-cipher.h
    ./include/sqlite3ext-cipher.h
    ./lib
    ./lib/libsqlite3-cipher.0.dylib
    ./lib/libsqlite3-cipher.a
    ./lib/libsqlite3-cipher.dylib
    ./lib/libsqlite3-cipher.la
    ./lib/pkgconfig
    ./lib/pkgconfig/sqlite3-cipher.pc

(On Mac OS X at least).

Programs can now use pkg-config on the name sqlite3-cipher to get the
new lib they are to link against, and instead of including sqlite3.h
they include sqlite3-cipher.h.

It is not possible to link against both sqlcipher and sqlite3 at the
same time, so if there are static libraries that are linked against
sqlite3 you can't link them into a project using sqlcipher eventhough
the name is different, they both still export the same exact names.

Ugh, what a mess.