Adds PRAGMA cipher_log_subsystem to restrict output of log messages
This change categorizes all sqlcipher_log calls with a subsystem of
either CORE, MEMORY, MUTEX, or PROVIDER. The pragma allows an application
to restrict logging of messages to a specific subsystem to reduce
the verbosity of log messages. This is most useful in the context of
TRACE and DEBUG level logging where the volume of log messages
is very high.