Adds private memory allocation and improvments to initialization and shutdown
This change introduces a private memory space for SQLCipher which is
allocated and locked once and they used to parcel out memory allocations
for parts of the library. In support of this objective the startup
and shutdown of the library have been completely reorganized to rely
on SQLite's SQLITE_EXTRA_INIT and SQLITE_EXTRA_SHUTDOWN.

History: In previous versions of SQLCipher small memory allocations were
made by various parts of the library as needed. Some systems limit the number
of memory pages that can be memlocked, however, and small allocations
caused undesirable fragmentation and locking issues. Recent changes to android
to reduce mlock quotas made it impossible to reliably lock memory segments.